Rewrite data fixtures logic, and remove function added for testing
