+++ tests/wycheproof/AbstractTestCase.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace Mdanter\Ecc\WycheProof;
 
+use Mdanter\Ecc\Crypto\Signature\HasherInterface;
+use Mdanter\Ecc\Crypto\Signature\SignHasher;
 use Mdanter\Ecc\Curves\NistCurve;
 use Mdanter\Ecc\Curves\SecgCurve;
 use Mdanter\Ecc\Math\MathAdapterFactory;
@@ -10,6 +12,28 @@ use PHPUnit\Framework\TestCase;
 
 abstract class AbstractTestCase extends TestCase
 {
+    protected $hashFunctions = [
+        "SHA-224" => "sha224",
+        "SHA-256" => "sha256",
+        "SHA-384" => "sha384",
+        "SHA-512" => "sha512",
+    ];
+
+    protected $curveAltName = [
+        "secp224r1" => NistCurve::NAME_P224,
+        "secp256r1" => NistCurve::NAME_P256,
+        "secp384r1" => NistCurve::NAME_P384,
+        "secp521r1" => NistCurve::NAME_P521,
+    ];
+
+    public function getHasher(string $id): HasherInterface
+    {
+        if (!array_key_exists($id, $this->hashFunctions)) {
+            throw new \RuntimeException("unconfigured hash function: $id");
+        }
+        return new SignHasher($this->hashFunctions[$id]);
+    }
+
     /**
      * @param array $extra
      * @return array
